+++ postgres-single/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Version bump for new base image
 * Fix node_exporter zfs issue
 * Update postgresql_exporter to 0.15.0
+* Add postgresql-contrib to support useful plugins like pg_stat_statements
 
 ---